requirements.txt
flask
line-bot-sdk
pandas
openpyxl
faiss-cpu
sentence-transformers
langchain
langchain-community
langchain-huggingface